+++ src/client/hooks/useThemaBreadcrumbs.ts
@@ -1,0 +1,27 @@
+import { type LinkProps, useLocation } from 'react-router';
+
+import { useThemaMenuItemByThemaID } from './useThemaMenuItems.ts';
+
+export function useThemaBreadcrumbs<ID extends string = string>(
+  themaID: ID
+): LinkProps[] {
+  const themaPaginaBreadcrumb = useThemaMenuItemByThemaID(themaID);
+  const location = useLocation();
+  const from = location?.state?.from;
+  const fromPageType = location?.state?.pageType;
+
+  return [
+    themaPaginaBreadcrumb
+      ? {
+          to: themaPaginaBreadcrumb?.to,
+          title: themaPaginaBreadcrumb?.title,
+        }
+      : null,
+    themaPaginaBreadcrumb && fromPageType === 'listpage'
+      ? {
+          to: from,
+          title: 'Lijst',
+        }
+      : null,
+  ].filter((link) => link !== null);
+}
