+++ src/client/pages/Thema/Afis/AfisFactuur.tsx
@@ -23,7 +23,7 @@ import { PageContentCell } from '../../../components/Page/Page.tsx';
 import { TableV2 } from '../../../components/Table/TableV2.tsx';
 import { ThemaDetailPagina } from '../../../components/Thema/ThemaDetailPagina.tsx';
 import { useHTMLDocumentTitle } from '../../../hooks/useHTMLDocumentTitle.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 function getTermijnenTable(factuur: AfisFactuurFrontend) {
   if (!factuur.termijnen || factuur.termijnen.length === 0) {
