+++ src/client/pages/Thema/Afis/useAfisListPageData.test.tsx
@@ -7,7 +7,7 @@ import { bffApi } from '../../../../testing/utils.ts';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
 
 vi.mock('../../../hooks/useAppStateStore');
-vi.mock('../../../hooks/useThemaMenuItems');
+vi.mock('../../../hooks/useThemaBreadcrumbs');
 
 const mockAFIS = {
   content: {
