+++ src/client/pages/Thema/Bezwaren/useBezwarenThemaData.hook.ts
@@ -3,7 +3,7 @@ import type { BezwaarFrontend } from '../../../../server/services/bezwaren/types
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useBezwarenThemaData() {
   const { BEZWAREN } = useAppStateGetter();
