+++ src/client/pages/Thema/Bodem/useBodemData.hook.tsx
@@ -4,7 +4,7 @@ import { tableConfig, themaConfig } from './Bodem-thema-config.ts';
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useBodemData() {
   const { BODEM } = useAppStateGetter();
