+++ src/client/pages/Thema/HLI/useHliThemaData.ts
@@ -17,7 +17,7 @@ import {
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAddDocumentLinkComponents } from '../../../data-transform/useAddDocumentLinks.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useHliThemaData() {
   const { HLI } = useAppStateGetter();
