+++ src/client/pages/Thema/Inkomen/useInkomenDetailData.hook.ts
@@ -3,7 +3,7 @@ import { useParams } from 'react-router';
 import { themaConfig } from './Inkomen-thema-config.ts';
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useInkomenDetailData(
   stateKey: 'WPI_TOZO' | 'WPI_BBZ' | 'WPI_AANVRAGEN' | 'WPI_TONK'
