+++ src/client/pages/Thema/Inkomen/useInkomenThemaData.hook.ts
@@ -14,7 +14,7 @@ import {
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAddDocumentLinkComponents } from '../../../data-transform/useAddDocumentLinks.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useInkomenThemaData() {
   const { WPI_AANVRAGEN, WPI_SPECIFICATIES, WPI_TOZO, WPI_TONK, WPI_BBZ } =
