+++ src/client/pages/Thema/Jeugd/useJeugdThemaData.ts
@@ -3,7 +3,7 @@ import { listPageParamKind, listPageTitle } from './Jeugd-thema-config.ts';
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useJeugdThemaData() {
   const { JEUGD } = useAppStateGetter();
