+++ src/client/pages/Thema/Klachten/useKlachtenThemaData.hook.ts
@@ -3,7 +3,7 @@ import type { KlachtFrontend } from '../../../../server/services/klachten/types.
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useKlachtenThemaData() {
   const { KLACHTEN } = useAppStateGetter();
