+++ src/client/pages/Thema/Parkeren/useParkerenData.hook.tsx
@@ -4,7 +4,7 @@ import type { DecosZaakFrontend } from '../../../../server/services/vergunningen
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useParkerenData() {
   const { PARKEREN } = useAppStateGetter();
