+++ src/client/pages/Thema/Profile/private/useWonenThemaData.hook.ts
@@ -1,6 +1,6 @@
 import { isError, isLoading } from '../../../../../universal/helpers/api.ts';
 import { useAppStateGetter } from '../../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../../hooks/useThemaBreadcrumbs.ts';
 import { themaConfig } from '../Profile-thema-config.ts';
 
 export function useWonenThemaData() {
