+++ src/client/pages/Thema/Varen/useVarenThemaData.hook.ts
@@ -7,7 +7,7 @@ import type {
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useVarenThemaData() {
   const { VAREN } = useAppStateGetter();
