+++ src/client/pages/Thema/Vergunningen/useVergunningenThemaData.hook.ts
@@ -3,7 +3,7 @@ import type { ZaakFrontendCombined } from '../../../../server/services/vergunnin
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useVergunningenThemaData() {
   const { VERGUNNINGEN, PARKEREN } = useAppStateGetter();
