+++ src/client/pages/Thema/Zorg/useZorgThemaData.ts
@@ -8,7 +8,7 @@ import type { WMOVoorzieningFrontend } from '../../../../server/services/jzd/wmo
 import { isError, isLoading } from '../../../../universal/helpers/api.ts';
 import { addLinkElementToProperty } from '../../../components/Table/TableV2.tsx';
 import { useAppStateGetter } from '../../../hooks/useAppStateStore.ts';
-import { useThemaBreadcrumbs } from '../../../hooks/useThemaMenuItems.ts';
+import { useThemaBreadcrumbs } from '../../../hooks/useThemaBreadcrumbs.ts';
 
 export function useZorgThemaData() {
   const { WMO } = useAppStateGetter();
